refactor: restructure modules
