+++ .gitignore
@@ -11,3 +11,5 @@ gradle-app.setting
 # gradle/wrapper/gradle-wrapper.properties
 
 .idea/uiDesigner.xml
+
+.idea/compiler.xml
