+++ .idea/modules/api/project-name.api.test.iml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="CheckStyle-IDEA-Module">
+    <option name="configuration">
+      <map />
+    </option>
+  </component>
+  <component name="TestModuleProperties" production-module="project-name.api.main" />
+</module>
