--- api/build.gradle
@@ -1,20 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-dependencies {
-    api 'com.vdurmont:semver4j:3.1.0'
-    api platform('org.springframework.boot:spring-boot-dependencies:2.6.0')
-    compileOnly 'org.springframework:spring-web'
-    compileOnly 'org.springframework:spring-webmvc'
-
-    compileOnly 'org.projectlombok:lombok:1.18.22'
-    annotationProcessor 'org.projectlombok:lombok:1.18.22'
-
-
-    compileOnly 'io.swagger.core.v3:swagger-annotations:2.1.11'
-    compileOnly 'javax.validation:validation-api:2.0.1.Final'
-}
-jacocoTestCoverageVerification {
-    enabled(false)
-}
