--- api/src/main/java/com/domain/projectname/models/VersionResponse.java
@@ -1,29 +1,0 @@
-package com.domain.projectname.models;
-
-import com.vdurmont.semver4j.Semver;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.validation.constraints.Pattern;
-
-import static lombok.AccessLevel.PRIVATE;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class VersionResponse {
-	@Pattern(regexp = "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$")
-	@Setter(PRIVATE)
-	private String  version;
-	@Setter(PRIVATE)
-	private boolean isStable;
-
-	public VersionResponse(Semver version) {
-		this(
-				version.getValue(),
-				version.isStable()
-		);
-	}
-}
