--- build.gradle
@@ -1,20 +1,0 @@
-plugins {
-    id 'org.sonarqube' version '3.3'
-}
-allprojects {
-    apply from: "${rootDir}/config.gradle"
-    version = '0.3.0'
-    group = project.ext.rootPackage
-}
-subprojects {
-    apply from: "${rootDir}/submodule.gradle"
-}
-sonarqube {
-    properties {
-        property 'sonar.projectName', project.ext.displayName
-        property "sonar.sourceEncoding", "UTF-8"
-        property "sonar.projectKey", project.ext.sonarProjectKey
-        property "sonar.organization", project.ext.sonarOrganization
-        property "sonar.host.url", project.ext.sonarHost
-    }
-}
