+++ modules/backend/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    compileOnly 'org.projectlombok:lombok:1.18.22'
+    annotationProcessor 'org.projectlombok:lombok:1.18.22'
+    compileOnly 'javax.validation:validation-api:2.0.1.Final'
+}
