+++ modules/backend/src/main/java/com/domain/projectname/models/VersionResponse.java
@@ -1,0 +1,16 @@
+package com.domain.projectname.models;
+
+import lombok.*;
+
+import javax.validation.constraints.Pattern;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class VersionResponse {
+	@Pattern(regexp = "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$")
+	@Setter(AccessLevel.PRIVATE)
+	private String  version;
+	@Setter(AccessLevel.PRIVATE)
+	private boolean isStable;
+}
