+++ settings.gradle
@@ -1,3 +1,8 @@
 rootProject.name = "project-name"
 
-include "app", "api", "backend"
+rootProject.setBuildFileName("modules/build.gradle")
+file('modules').eachDir { dir ->
+    println "Add module $dir"
+    include dir.name
+    project(":${dir.name}").projectDir = dir
+}
